+++ dotnet/Base/Repository/Domain/Base/Content/FileMediaContent.cs
@@ -1,0 +1,41 @@
+// <copyright file="FileMediaContent.cs" company="Allors bv">
+// Copyright (c) Allors bv. All rights reserved.
+// Licensed under the LGPL license. See LICENSE file in the project root for full license information.
+// </copyright>
+// <summary>Defines the FileMediaContent type.</summary>
+
+namespace Allors.Repository
+{
+    using Attributes;
+    using static Workspaces;
+
+
+    #region Allors
+    [Id("b06e49e2-4aac-4454-ad6a-68ce5010b5e2")]
+    #endregion
+    [Workspace(Default)]
+    public partial class FileMediaContent : MediaContent
+    {
+        #region inherited properties
+        public Revocation[] Revocations { get; set; }
+
+        public SecurityToken[] SecurityTokens { get; set; }
+
+        public string Type { get; set; }
+        #endregion
+
+        #region inherited methods
+        public void OnBuild() { }
+
+        public void OnPostBuild() { }
+
+        public void OnInit()
+        {
+        }
+
+        public void OnPostDerive() { }
+
+        public void Delete() { }
+        #endregion
+    }
+}